refacroting types and folders
