+++ src/components/DescriptionFilm/DescriptionFilm.tsx
@@ -1,22 +1,19 @@
-/* eslint-disable no-console */
 import { useContext } from "react";
-// import { useAppDispatch } from "../../hooks/hook";
-import { MyDescriptionFilmContext } from "../../pages/Film";
-// import { addToFavorites } from "../../store/favoritesSlice";
-import s from "./DescriptionFilm.module.css";
-import { useFavorites } from "../../hooks/useFavorites";
 import { getAuth } from "firebase/auth";
 import { auth } from "../../config/firebase-config";
 
+import { useFavorites } from "../../hooks/useFavorites";
+
+import { MyDescriptionFilmContext } from "../../pages/FilmPage";
+
+import s from "./DescriptionFilm.module.css";
+
 export const DescriptionFilm = () => {
-  const { description, nameRu, ratingKinopoisk, year, posterUrl, id } = useContext(
-    MyDescriptionFilmContext
-  );
+  const { description, nameRu, ratingKinopoisk, year, posterUrl, id } =
+    useContext(MyDescriptionFilmContext);
   const isAuth = getAuth();
   const userId = auth.currentUser?.uid;
   const userEmail = isAuth.currentUser?.email;
-
-  // const dispatch = useAppDispatch();
 
   const { addToFavorites } = useFavorites(`${userEmail}`);
 
@@ -28,20 +25,11 @@ export const DescriptionFilm = () => {
         posterUrl: posterUrl,
         ratingKinopoisk: ratingKinopoisk,
         year: year,
-        id: id
+        id: id,
       });
     }
-    // addToFavorites({
-    //   description: description!,
-    //   nameRu: nameRu!,
-    //   posterUrl: posterUrl!,
-    //   ratingKinopoisk: ratingKinopoisk!,
-    //   year: year!,
-    // })
-
-    //   )
-    // }
   };
+
   return (
     <div className={s["film-description"]}>
       <h2>{nameRu}</h2>
