+++ src/components/FilmSearchResults/FilmSearchResult.props.ts
@@ -1,0 +1,4 @@
+
+export interface FilmSearchResultsInterface {
+    searchTerm: string;
+  }
