+++ src/components/Input/Input.tsx
@@ -1,7 +1,9 @@
 import { forwardRef } from "react";
-import s from "./Input.module.css";
 import cn from "classnames";
+
 import { InputProps } from "./Input.props";
+
+import s from "./Input.module.css";
 
 export const Input = forwardRef<HTMLInputElement, InputProps>(function Input(
   { isValid = true, className, ...props },
@@ -17,5 +19,3 @@ export const Input = forwardRef<HTMLInputElement, InputProps>(function Input(
     />
   );
 });
-
-
