+++ src/components/RegistartionGroup/RegistrationGroup.tsx
@@ -1,5 +1,5 @@
 import { useNavigate } from "react-router-dom";
-import Button from "../Button/Button";
+import { Button } from "../Button/Button";
 
 import s from "./RegistrationGroup.module.css";
