--- src/components/Wrapper/Wrapper.module.css
@@ -1,9 +1,0 @@
-.wrapper {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-wrap: wrap;
-    gap: 40px;
-    padding: 20px 0px;
-  }
-
