--- src/components/Wrapper/Wrapper.props.ts
@@ -1,5 +1,0 @@
-import { HTMLAttributes, ReactNode } from "react";
-
-export interface WrapperProps extends HTMLAttributes<HTMLDivElement> {
-  children: ReactNode;
-}
