+++ src/hooks/index.ts
@@ -1,5 +1,5 @@
 export * from "./hook";
 export * from "./useDebounce";
 export * from "./useFavorites";
-// export * from "./useHistory";
+export * from "./useHistory";
 export * from "./useUserAuth";
