--- src/pages/Error/Error.tsx
@@ -1,3 +1,0 @@
-export function Error() {
-  return <>Error</>;
-}
