+++ src/pages/ErrorPage/ErrorPage.tsx
@@ -1,0 +1,3 @@
+export const ErrorPage = () => {
+  return <>Error</>;
+};
