+++ src/pages/FavoritesPage/FavoritesPage.module.css
