+++ src/pages/FilmsPage/FilmList/FilmList.tsx
@@ -1,7 +1,7 @@
-/* eslint-disable no-console */
-import s from "./FilmList.module.css";
 import { FilmCart } from "../../../components/FilmCart/FilmCart";
 import { FilmListPropsType } from "./FilmList.props";
+
+import s from "./FilmList.module.css";
 
 export const FilmList = ({ data }: FilmListPropsType) => {
   return (
@@ -11,7 +11,7 @@ export const FilmList = ({ data }: FilmListPropsType) => {
           key={item.kinopoiskId}
           id={item.kinopoiskId}
           title={item.nameOriginal || item.nameRu}
-          description={item.type}
+          // description={item.type}
           name={item.nameRu}
           rating={item.ratingImdb || item.ratingKinopoisk}
           image={item.posterUrlPreview}
