--- src/pages/History/index.ts
@@ -1,1 +1,0 @@
-export * from "./History";
