+++ src/pages/index.ts
@@ -1,8 +1,8 @@
-export * from './Error';
-export * from './Favorites';
-export * from './Film';
-export * from './FilmsPage';
-export * from './History';
-export * from './Login';
-export * from './Register';
-export * from './SearchPage';
+export * from "./ErrorPage";
+export * from "./FavoritesPage";
+export * from "./FilmPage";
+export * from "./FilmsPage";
+export * from "./HistoryPage";
+export * from "./LoginPage";
+export * from "./RegisterPage";
+export * from "./SearchPage";
