+++ src/store/favoritesSlice.ts
@@ -1,16 +1,9 @@
 /* eslint-disable no-console */
 import { createSlice } from "@reduxjs/toolkit";
+import { FavoritesFilmInterface } from "./types";
 
-interface FavoritesFilmInterface {
-  posterUrl: string;
-  description: string;
-  nameRu: string;
-  ratingKinopoisk: number;
-  year: number;
-}
-export type FavoritesState = {
-  films: FavoritesFilmInterface[];
-};
+
+
 
 const initialState: FavoritesFilmInterface[] = [
   {
