+++ src/store/types/favotites-type.ts
@@ -1,0 +1,10 @@
+export interface FavoritesFilmInterface {
+    posterUrl: string;
+    description: string;
+    nameRu: string;
+    ratingKinopoisk: number;
+    year: number;
+  }
+  export type FavoritesState = {
+    films: FavoritesFilmInterface[];
+  };
