+++ src/store/types/user-types.ts
@@ -1,0 +1,9 @@
+interface UserInterface {
+    email: string;
+    uid: string;
+  }
+  
+  export type UserState = {
+    authorize: boolean;
+    user: UserInterface | null;
+  };
