+++ src/utils/utils-types.ts
@@ -1,0 +1,4 @@
+export interface UserContextInterface {
+    email: string;
+    uid: string;
+  }
